telehealth_connect/requirements.txt
# TeleHealth Connect - Python Dependencies

# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3

# Database
PyMySQL==1.1.0
cryptography>=45.0.7,<47

# Password Hashing
Werkzeug==3.0.1

# Environment Variables
python-dotenv==1.0.0

# Development Dependencies (optional)
python-dateutil==2.8.2
